+++ src/components/page-content/photo/PhotoPageHeader.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
-import AppHeader from "../../components/layout/AppHeader/AppHeader.vue";
-import IconButton from "../../components/ui/IconButton.vue";
-import LikeButton from "../../components/photo/like-button/LikeButton.vue";
-import { LIKE_BUTTON_VARIANTS } from "../../components/photo/like-button/types";
-import { usePhotoStore } from "../../stores/photo";
+import AppHeader from "../../layout/AppHeader/AppHeader.vue";
+import IconButton from "../../ui/IconButton.vue";
+import LikeButton from "../../photo/like-button/LikeButton.vue";
+import { LIKE_BUTTON_VARIANTS } from "../../photo/like-button/types";
+import { usePhotoStore } from "../../../stores/photo";
 import { storeToRefs } from "pinia";
-import { PAGE_NAMES } from "../../common/constants/router";
+import { PAGE_NAMES } from "../../../common/constants/router";
 
 const store = usePhotoStore();
 const { isLiked, likeProcess } = storeToRefs(store);
